# swagger.json
{
	"components": {
...
		"schemas": {
			"ProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"lastOnline": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/UserDto"
					}
				},
				"required": [
					"id",
					"userId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ERole": {
				"enum": [
					"admin",
					"user",
					"guest"
				],
				"type": "string"
			},
...
			"IPermissionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EPermissions"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/ERole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"profile": {
						"$ref": "#/components/schemas/ProfileDto"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/IRoleDto"
						},
						"type": "array"
					},
					"directPermissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"roles",
					"directPermissions",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ITokensDto": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserWithTokensDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"profile": {
						"$ref": "#/components/schemas/ProfileDto"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/IRoleDto"
						},
						"type": "array"
					},
					"directPermissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"tokens": {
						"$ref": "#/components/schemas/ITokensDto"
					}
				},
				"required": [
					"id",
					"username",
					"roles",
					"directPermissions",
					"createdAt",
					"updatedAt",
					"tokens"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSignUpRequestDto": {
				"allOf": [
					{
						"properties": {
							"password": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							}
						},
						"required": [
							"password"
						],
						"type": "object"
					},
					{
						"anyOf": [
							{
								"properties": {
									"phone": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"phone"
								],
								"type": "object"
							},
							{
								"properties": {
									"phone": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						]
					}
				]
			},
			"I2FARequiredDto": {
				"properties": {
					"require2FA": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"twoFactorToken": {
						"type": "string"
					},
					"twoFactorHint": {
						"type": "string"
					}
				},
				"required": [
					"require2FA",
					"twoFactorToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignInResponseDto": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IUserWithTokensDto"
					},
					{
						"$ref": "#/components/schemas/I2FARequiredDto"
					}
				]
			},
			"ISignInRequestDto": {
				"properties": {
					"login": {
						"type": "string",
						"description": "Может быть телефоном, email-ом и username-ом"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"login",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserLoginRequestDto": {
				"properties": {
					"login": {
						"type": "string",
						"description": "Может быть телефоном, email-ом и username-ом"
					}
				},
				"required": [
					"login"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserResetPasswordRequestDto": {
				"properties": {
					"password": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEnable2FARequestDto": {
				"properties": {
					"password": {
						"type": "string"
					},
					"hint": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDisable2FARequestDto": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerify2FARequestDto": {
				"properties": {
					"twoFactorToken": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"twoFactorToken",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EMessageType": {
				"enum": [
					"text",
					"image",
					"file",
					"voice",
					"system",
					"poll",
					"sticker"
				],
				"type": "string"
			},
			"EMessageStatus": {
				"enum": [
					"sent",
					"delivered",
					"read"
				],
				"type": "string"
			},
			"Record_string.unknown_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"MessageDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"chatId": {
						"type": "string"
					},
					"senderId": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EMessageType"
					},
					"status": {
						"$ref": "#/components/schemas/EMessageStatus"
					},
					"content": {
						"type": "string",
						"nullable": true
					},
					"replyToId": {
						"type": "string",
						"nullable": true
					},
					"forwardedFromId": {
						"type": "string",
						"nullable": true
					},
					"isEdited": {
						"type": "boolean"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"isPinned": {
						"type": "boolean"
					},
					"pinnedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"pinnedById": {
						"type": "string",
						"nullable": true
					},
					"stickerId": {
						"type": "string",
						"nullable": true
					},
					"encryptedContent": {
						"type": "string",
						"nullable": true
					},
					"encryptionMetadata": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.unknown_"
							}
						],
						"nullable": true
					},
					"keyboard": {
						"nullable": true
					},
					"linkPreviews": {
						"items": {
							"properties": {
								"siteName": {
									"type": "string",
									"nullable": true
								},
								"imageUrl": {
									"type": "string",
									"nullable": true
								},
								"description": {
									"type": "string",
									"nullable": true
								},
								"title": {
									"type": "string",
									"nullable": true
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"siteName",
								"imageUrl",
								"description",
								"title",
								"url"
							],
							"type": "object"
						},
						"type": "array",
						"nullable": true
					},
					"scheduledAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"isScheduled": {
						"type": "boolean"
					},
					"selfDestructSeconds": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"selfDestructAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"sender": {
						"properties": {
							"avatarUrl": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"replyTo": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MessageDto"
							}
						],
						"nullable": true
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/MessageAttachmentDto"
						},
						"type": "array"
					},
					"reactions": {
						"items": {
							"properties": {
								"userIds": {
									"items": {
										"type": "string"
									},
									"type": "array"
								},
								"count": {
									"type": "number",
									"format": "double"
								},
								"emoji": {
									"type": "string"
								}
							},
							"required": [
								"userIds",
								"count",
								"emoji"
							],
							"type": "object"
						},
						"type": "array"
					},
					"mentions": {
						"items": {
							"properties": {
								"isAll": {
									"type": "boolean"
								},
								"userId": {
									"type": "string",
									"nullable": true
								}
							},
							"required": [
								"isAll",
								"userId"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"chatId",
					"senderId",
					"type",
					"status",
					"content",
					"replyToId",
					"forwardedFromId",
					"isEdited",
					"isDeleted",
					"isPinned",
					"pinnedAt",
					"pinnedById",
					"stickerId",
					"encryptedContent",
					"encryptionMetadata",
					"keyboard",
					"linkPreviews",
					"scheduledAt",
					"isScheduled",
					"selfDestructSeconds",
					"selfDestructAt",
					"createdAt",
					"updatedAt",
					"attachments",
					"reactions",
					"mentions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageAttachmentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fileId": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"fileType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					},
					"thumbnailUrl": {
						"type": "string",
						"nullable": true
					},
					"width": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"height": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"id",
					"fileId",
					"fileName",
					"fileUrl",
					"fileType",
					"fileSize",
					"thumbnailUrl",
					"width",
					"height"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBotSendMessageBody": {
				"properties": {
					"chatId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EMessageType"
					},
					"replyToId": {
						"type": "string"
					},
					"fileIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"chatId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBotEditMessageBody": {
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ISendMessageBody": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/EMessageType"
					},
					"content": {
						"type": "string"
					},
					"replyToId": {
						"type": "string"
					},
					"forwardedFromId": {
						"type": "string"
					},
					"fileIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mentionedUserIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mentionAll": {
						"type": "boolean"
					},
					"encryptedContent": {
						"type": "string"
					},
					"encryptionMetadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"scheduledAt": {
						"type": "string"
					},
					"selfDestructSeconds": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IMessageListDto": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MessageDto"
						},
						"type": "array"
					},
					"hasMore": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"hasMore"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IEditMessageBody": {
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.1",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/auth/sign-up": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithTokensDto"
								}
							}
						}
					}
				},
				"description": "Регистрация нового пользователя",
				"summary": "Регистрация",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Данные для регистрации",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TSignUpRequestDto",
								"description": "Данные для регистрации"
							},
							"example": {
								"email": "user@example.com",
								"password": "password123",
								"firstName": "John",
								"lastName": "Doe"
							}
						}
					}
				}
			}
		},
		"/api/auth/sign-in": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISignInResponseDto"
								}
							}
						}
					}
				},
				"description": "Авторизация пользователя",
				"summary": "Вход в систему",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Данные для входа",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignInRequestDto",
								"description": "Данные для входа"
							},
							"example": {
								"login": "epifanovmd@gmail.com",
								"password": "Epifan123"
							}
						}
					}
				}
			}
		},
		"/api/auth/request-reset-password": {
			"post": {
				"operationId": "RequestResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"description": "Запрос на сброс пароля",
				"summary": "Запрос сброса пароля",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserLoginRequestDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"description": "Сброс пароля",
				"summary": "Смена пароля",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Токен и новый пароль",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserResetPasswordRequestDto",
								"description": "Токен и новый пароль"
							},
							"example": {
								"token": "reset-token-123",
								"password": "newSecurePassword"
							}
						}
					}
				}
			}
		},
		"/api/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITokensDto"
								}
							}
						}
					}
				},
				"description": "Обновление токенов доступа",
				"summary": "Обновление токенов",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Тело запроса с refresh токеном",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object",
								"description": "Тело запроса с refresh токеном"
							},
							"example": {
								"refreshToken": "old-refresh-token"
							}
						}
					}
				}
			}
		},
		"/api/auth/enable-2fa": {
			"post": {
				"operationId": "Enable2FA",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"description": "Включить двухфакторную аутентификацию.",
				"summary": "Включение 2FA",
				"tags": [
					"Authorization"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEnable2FARequestDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/disable-2fa": {
			"post": {
				"operationId": "Disable2FA",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"description": "Отключить двухфакторную аутентификацию.",
				"summary": "Отключение 2FA",
				"tags": [
					"Authorization"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDisable2FARequestDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/verify-2fa": {
			"post": {
				"operationId": "Verify2FA",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithTokensDto"
								}
							}
						}
					}
				},
				"description": "Верифицировать 2FA и получить токены.",
				"summary": "Верификация 2FA",
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerify2FARequestDto"
							}
						}
					}
				}
			}
		},
...
		"/api/bot-api/message/send": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Отправить сообщение от имени бота.",
				"summary": "Bot: отправка сообщения",
				"tags": [
					"Bot API"
				],
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBotSendMessageBody"
							}
						}
					}
				}
			}
		},
		"/api/bot-api/message/{id}/edit": {
			"post": {
				"operationId": "EditMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Редактировать сообщение бота.",
				"summary": "Bot: редактирование сообщения",
				"tags": [
					"Bot API"
				],
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBotEditMessageBody"
							}
						}
					}
				}
			}
		},
		"/api/bot-api/message/{id}": {
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Удалить сообщение бота.",
				"summary": "Bot: удаление сообщения",
				"tags": [
					"Bot API"
				],
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/chat/{chatId}/message": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Отправить сообщение в чат.",
				"summary": "Отправка сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISendMessageBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessageListDto"
								}
							}
						}
					}
				},
				"description": "Получить сообщения чата с cursor-based пагинацией.",
				"summary": "Список сообщений",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID чата",
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "ID сообщения для курсора (загрузить более старые)",
						"in": "query",
						"name": "before",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Количество сообщений (по умолчанию 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/message/{id}": {
			"patch": {
				"operationId": "EditMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Отредактировать сообщение.",
				"summary": "Редактирование сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEditMessageBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Удалить сообщение (soft delete).",
				"summary": "Удаление сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		},
		{
			"url": "http://0.0.0.0:8181",
			"description": "Local dev"
		},
		{
			"url": "http://147.45.245.104:8181",
			"description": "Host"
		}
	]
}